Sentry Integration - Top level unhandle exception logging - TypeScript support with RewriteFrames config
Moving the sourcemappingurl to script so it can be handled in GH Actions
Sentry upload moved after changeset version bump script and npm script to get current version into GH env variable
add org and project to secrets

Prompt for Opt in to Sentry
- Check occurs on the use of Wrangler, initial use when no property for Sentry is
present it will prompt the user if they would prefer to opt in to using Sentry.
- If the property already exists it will skip the prompt, if true it will initialize Sentry, if false it will
disable Sentry so that it cannot send captured exceptions

Sentry Tests:wq
- Tests surfaced that the permissions needed to to check if the file existed
if the default.toml did not exist it exits the function returning undefined which prompts the
user then creates default.toml if necessary and adding the property to save the user decision.
- The tests currently check that the decision flow works as currently set up then checks if Sentry is able
to send events or is disabled.